Extract Dashboard JITM host
